+++ Components/SwatchColor.cs
@@ -1,0 +1,8 @@
+﻿namespace WebAwesomeBlazor.Components
+{
+    public class SwatchColor
+    {
+        public string Color { get; set; } = default!;
+        public string? Label { get; set; }
+    }
+}